sr(api): move update-training in devcomp
